UDP,TCP fix: port numbers were represented with (signed) short at places; SCTP: minor related change
Signed short caused problems for large port numbers due to sign
extension (see ASSERT(conn->localPort == localPort) in TCP.cc)
[reported by Thomas Gamer]

+++ src/transport/sctp/SCTPAssociationRcvMessage.cc
@@ -71,8 +71,8 @@ bool SCTPAssociation::process_RCV_Message(SCTPMessage*       sctpmsg,
     }
 
     SCTPPathVariables* path      = getPath(src);
-    const int32  srcPort             = sctpmsg->getDestPort();
-    const int32  destPort        = sctpmsg->getSrcPort();
+    const uint16 srcPort         = sctpmsg->getDestPort();
+    const uint16 destPort        = sctpmsg->getSrcPort();
     const uint32 numberOfChunks = sctpmsg->getChunksArraySize();
     sctpEV3 << "numberOfChunks=" <<numberOfChunks << endl;
